# api/openapi/admin-domains.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/seo/opportunities/{id}/brief": {
      "post": {
        "operationId": "SeoAutomationController_createBrief",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
...
    "/api/admin/seo/recommendations/{id}/apply": {
      "post": {
        "operationId": "SeoAutomationController_applyRecommendation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
    "/api/admin/seo/recommendations/{id}/reject": {
      "post": {
        "operationId": "SeoAutomationController_rejectRecommendation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
...
    "/api/admin/seo/content-pipeline/{id}/publish": {
      "post": {
        "operationId": "SeoAutomationController_publishBrief",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
...
    "/api/admin/seo/internal-links/{id}/apply": {
      "post": {
        "operationId": "SeoAutomationController_applyInternalLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
    "/api/admin/seo/change-log": {
      "get": {
        "operationId": "SeoAutomationController_getChangeLog",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
    "/api/admin/seo/products/{id}/seo/generate": {
      "post": {
        "operationId": "SeoAutomationController_generateProductSeoDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
    "/api/admin/seo/products/{id}/seo/apply": {
      "post": {
        "operationId": "SeoAutomationController_applyProductSeoDraft",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "admin-seo-automation"
        ]
      }
    },
...
  },
  "info": {
    "title": "PulseGear Admin Domains API",
    "description": "OpenAPI schema for admin product-research and seo-automation domains",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}